requirements.txt
scikit-learn
albumentations>=1.1.0
efficientnet_pytorch>=0.7.1
matplotlib>=3.2.2
numpy>=1.21.2
opencv_python_headless>=4.5.5.62
prefetch_generator>=1.0.1
pretrainedmodels>=0.7.4
PyYAML>=3.13
scipy>=1.4.1
seaborn>=0.11.2
tensorboardX>=2.4.1
timm==0.5.4
torch>=1.10.0
torchvision>=0.11.0
tqdm>=4.61.2
webcolors>=1.11.1
urllib3==1.26.6 
Flask
modin[all]
opencv-python
scikit-image
pillow
pixellib
yolov5
streamlit
streamlit-webrtc==0.36.0
tensorflow
tflite_runtime